fix: Check for field choices before building enum for select fields (#31)
* Check for field choices before building enum for selects

When converting an Airtable type into a Zod type in
`airtableTypeToZodType(...)`, there are a few checks against sets of types that
reduce to a single Zod type (`STRING_TYPES`, `NUMBER_TYPES`, etc.). After these
checks, the function builds a list of the choices for a field.

If the `options.choices` member is undefined, the call to `map()` results in an
exception being thrown. For single- and multi-select fields, the function
continues, but for any types in the `TYPE_MAP` object, the function will fail
(except for select fields, which will have string values).

This change wraps the code for strongly-typed select fields by first checking
if the `options.choices` member exists. If it does, it builds the choices list
and checks for a select field. Otherwise, it continues to consulting
`TYPE_MAP`.

* Create hungry-bikes-doubt.md

---------

Co-authored-by: Matt Kane <m@mk.gg>